Add/modify new collaborator component added
